+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 0.4.5 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Management command to poll provider status changes.
 
 
 0.4.4 (2019-01-25)